Dataset and analysis code
Added the csv file for the CO2 emissions from 1750-2022 and the MATLAB code to get the statistics for each country. If you need to run the code yourself, make sure you download the csv file and copy the file path from your computer and enter it into the MATLAB code on line 5.