Kyle Pruitt - Final Code and Dataset (#11)
* Kyle Pruitt - Update README.md

Added my role and contribution to the readme file.

* MATLAB analysis code draft1

Initial MATLAB code for analyzing the CO2 emissions by country from 1750 to 2022. If you are going to run this on your computer you need to download the data set and then input the copy path file into the read part of the code. Provides the count, mean, standard deviation, minimum, median, maximum, and sum for each country and variable. Needs adjusting to reduce the data down to just the CO2 emissions. Some of the statistics are thrown off such as the standard deviation and mean because most countries had zero values for most values before the year 1900. Need to discuss what exactly we want out of this data and then adjusting the code accordingly.

* Dataset and analysis code

Added the csv file for the CO2 emissions from 1750-2022 and the MATLAB code to get the statistics for each country. If you need to run the code yourself, make sure you download the csv file and copy the file path from your computer and enter it into the MATLAB code on line 5.

* Deleted analysis code draft